refactor: abstract move to trash note hook
